Use Poetry for dependency management and other initial work (#21)
This PR wraps up some initial work needed to better structure the
project.

* Use Poetry for dependency management.
* Refactor codebase. Resolves #15.
* Update .gitignore and remove unnecessary files. Resolves #16.
* Save GBRT model using skops instead of joblib. Resolves #19.

---------
