--- .VSCodeCounter/2023-03-27_11-43-01/details.md
@@ -1,27 +1,0 @@
-# Details
-
-Date : 2023-03-27 11:43:01
-
-Directory /Users/gbastien/Documents/Research/Projects/VirusInfectionPredictor/vip/
-
-Total : 12 files,  1093 codes, 39 comments, 249 blanks, all 1381 lines
-
-[Summary](results.md) / Details / [Diff Summary](diff.md) / [Diff Details](diff-details.md)
-
-## Files
-| filename | language | code | comment | blank | total |
-| :--- | :--- | ---: | ---: | ---: | ---: |
-| [vip/__init__.py](/vip/__init__.py) | Python | 0 | 0 | 1 | 1 |
-| [vip/mlmodel/README.md](/vip/mlmodel/README.md) | Markdown | 0 | 0 | 1 | 1 |
-| [vip/mlmodel/__init__.py](/vip/mlmodel/__init__.py) | Python | 0 | 0 | 1 | 1 |
-| [vip/mlmodel/build_model.ipynb](/vip/mlmodel/build_model.ipynb) | JSON | 509 | 0 | 1 | 510 |
-| [vip/mlmodel/compute_ml_features.py](/vip/mlmodel/compute_ml_features.py) | Python | 265 | 9 | 86 | 360 |
-| [vip/mlmodel/genes_features.py](/vip/mlmodel/genes_features.py) | Python | 2 | 1 | 12 | 15 |
-| [vip/mlmodel/genomes_features.py](/vip/mlmodel/genomes_features.py) | Python | 189 | 17 | 68 | 274 |
-| [vip/mlmodel/hyperparameters.py](/vip/mlmodel/hyperparameters.py) | Python | 0 | 0 | 1 | 1 |
-| [vip/mlmodel/read_sequence.py](/vip/mlmodel/read_sequence.py) | Python | 15 | 0 | 6 | 21 |
-| [vip/network/gephi_input.py](/vip/network/gephi_input.py) | Python | 0 | 0 | 1 | 1 |
-| [vip/network/properties.py](/vip/network/properties.py) | Python | 70 | 8 | 47 | 125 |
-| [vip/predict_interactions.py](/vip/predict_interactions.py) | Python | 43 | 4 | 24 | 71 |
-
-[Summary](results.md) / Details / [Diff Summary](diff.md) / [Diff Details](diff-details.md)
