--- .VSCodeCounter/2023-03-27_11-43-01/diff-details.md
@@ -1,15 +1,0 @@
-# Diff Details
-
-Date : 2023-03-27 11:43:01
-
-Directory /Users/gbastien/Documents/Research/Projects/VirusInfectionPredictor/vip/
-
-Total : 0 files,  0 codes, 0 comments, 0 blanks, all 0 lines
-
-[Summary](results.md) / [Details](details.md) / [Diff Summary](diff.md) / Diff Details
-
-## Files
-| filename | language | code | comment | blank | total |
-| :--- | :--- | ---: | ---: | ---: | ---: |
-
-[Summary](results.md) / [Details](details.md) / [Diff Summary](diff.md) / Diff Details
