--- .VSCodeCounter/2023-03-27_11-43-01/diff.csv
@@ -1,2 +1,0 @@
-"filename", "language", "", "comment", "blank", "total"
-"Total", "-", , 0, 0, 0
