--- .VSCodeCounter/2023-03-27_11-43-01/diff.md
@@ -1,19 +1,0 @@
-# Diff Summary
-
-Date : 2023-03-27 11:43:01
-
-Directory /Users/gbastien/Documents/Research/Projects/VirusInfectionPredictor/vip/
-
-Total : 0 files,  0 codes, 0 comments, 0 blanks, all 0 lines
-
-[Summary](results.md) / [Details](details.md) / Diff Summary / [Diff Details](diff-details.md)
-
-## Languages
-| language | files | code | comment | blank | total |
-| :--- | ---: | ---: | ---: | ---: | ---: |
-
-## Directories
-| path | files | code | comment | blank | total |
-| :--- | ---: | ---: | ---: | ---: | ---: |
-
-[Summary](results.md) / [Details](details.md) / Diff Summary / [Diff Details](diff-details.md)
