--- .VSCodeCounter/2023-03-27_11-43-01/diff.txt
@@ -1,22 +1,0 @@
-Date : 2023-03-27 11:43:01
-Directory : /Users/gbastien/Documents/Research/Projects/VirusInfectionPredictor/vip/
-Total : 0 files,  0 codes, 0 comments, 0 blanks, all 0 lines
-
-Languages
-+----------+------------+------------+------------+------------+------------+
-| language | files      | code       | comment    | blank      | total      |
-+----------+------------+------------+------------+------------+------------+
-+----------+------------+------------+------------+------------+------------+
-
-Directories
-+------+------------+------------+------------+------------+------------+
-| path | files      | code       | comment    | blank      | total      |
-+------+------------+------------+------------+------------+------------+
-+------+------------+------------+------------+------------+------------+
-
-Files
-+----------+----------+------------+------------+------------+------------+
-| filename | language | code       | comment    | blank      | total      |
-+----------+----------+------------+------------+------------+------------+
-| Total    |          |          0 |          0 |          0 |          0 |
-+----------+----------+------------+------------+------------+------------+
