--- .VSCodeCounter/2023-03-27_11-43-01/results.csv
@@ -1,14 +1,0 @@
-"filename", "language", "Python", "Markdown", "JSON", "comment", "blank", "total"
-"/Users/gbastien/Documents/Research/Projects/VirusInfectionPredictor/vip/__init__.py", "Python", 0, 0, 0, 0, 1, 1
-"/Users/gbastien/Documents/Research/Projects/VirusInfectionPredictor/vip/mlmodel/README.md", "Markdown", 0, 0, 0, 0, 1, 1
-"/Users/gbastien/Documents/Research/Projects/VirusInfectionPredictor/vip/mlmodel/__init__.py", "Python", 0, 0, 0, 0, 1, 1
-"/Users/gbastien/Documents/Research/Projects/VirusInfectionPredictor/vip/mlmodel/build_model.ipynb", "JSON", 0, 0, 509, 0, 1, 510
-"/Users/gbastien/Documents/Research/Projects/VirusInfectionPredictor/vip/mlmodel/compute_ml_features.py", "Python", 265, 0, 0, 9, 86, 360
-"/Users/gbastien/Documents/Research/Projects/VirusInfectionPredictor/vip/mlmodel/genes_features.py", "Python", 2, 0, 0, 1, 12, 15
-"/Users/gbastien/Documents/Research/Projects/VirusInfectionPredictor/vip/mlmodel/genomes_features.py", "Python", 189, 0, 0, 17, 68, 274
-"/Users/gbastien/Documents/Research/Projects/VirusInfectionPredictor/vip/mlmodel/hyperparameters.py", "Python", 0, 0, 0, 0, 1, 1
-"/Users/gbastien/Documents/Research/Projects/VirusInfectionPredictor/vip/mlmodel/read_sequence.py", "Python", 15, 0, 0, 0, 6, 21
-"/Users/gbastien/Documents/Research/Projects/VirusInfectionPredictor/vip/network/gephi_input.py", "Python", 0, 0, 0, 0, 1, 1
-"/Users/gbastien/Documents/Research/Projects/VirusInfectionPredictor/vip/network/properties.py", "Python", 70, 0, 0, 8, 47, 125
-"/Users/gbastien/Documents/Research/Projects/VirusInfectionPredictor/vip/predict_interactions.py", "Python", 43, 0, 0, 4, 24, 71
-"Total", "-", 584, 0, 509, 39, 249, 1381
