--- .VSCodeCounter/2023-03-27_11-43-01/results.json
@@ -1,1 +1,0 @@
-{"file:///Users/gbastien/Documents/Research/Projects/VirusInfectionPredictor/vip/__init__.py":{"language":"Python","code":0,"comment":0,"blank":1},"file:///Users/gbastien/Documents/Research/Projects/VirusInfectionPredictor/vip/network/gephi_input.py":{"language":"Python","code":0,"comment":0,"blank":1},"file:///Users/gbastien/Documents/Research/Projects/VirusInfectionPredictor/vip/network/properties.py":{"language":"Python","code":70,"comment":8,"blank":47},"file:///Users/gbastien/Documents/Research/Projects/VirusInfectionPredictor/vip/predict_interactions.py":{"language":"Python","code":43,"comment":4,"blank":24},"file:///Users/gbastien/Documents/Research/Projects/VirusInfectionPredictor/vip/mlmodel/README.md":{"language":"Markdown","code":0,"comment":0,"blank":1},"file:///Users/gbastien/Documents/Research/Projects/VirusInfectionPredictor/vip/mlmodel/__init__.py":{"language":"Python","code":0,"comment":0,"blank":1},"file:///Users/gbastien/Documents/Research/Projects/VirusInfectionPredictor/vip/mlmodel/hyperparameters.py":{"language":"Python","code":0,"comment":0,"blank":1},"file:///Users/gbastien/Documents/Research/Projects/VirusInfectionPredictor/vip/mlmodel/compute_ml_features.py":{"language":"Python","code":265,"comment":9,"blank":86},"file:///Users/gbastien/Documents/Research/Projects/VirusInfectionPredictor/vip/mlmodel/genes_features.py":{"language":"Python","code":2,"comment":1,"blank":12},"file:///Users/gbastien/Documents/Research/Projects/VirusInfectionPredictor/vip/mlmodel/build_model.ipynb":{"language":"JSON","code":509,"comment":0,"blank":1},"file:///Users/gbastien/Documents/Research/Projects/VirusInfectionPredictor/vip/mlmodel/genomes_features.py":{"language":"Python","code":189,"comment":17,"blank":68},"file:///Users/gbastien/Documents/Research/Projects/VirusInfectionPredictor/vip/mlmodel/read_sequence.py":{"language":"Python","code":15,"comment":0,"blank":6}}
