--- .VSCodeCounter/2023-03-27_11-43-01/results.md
@@ -1,26 +1,0 @@
-# Summary
-
-Date : 2023-03-27 11:43:01
-
-Directory /Users/gbastien/Documents/Research/Projects/VirusInfectionPredictor/vip/
-
-Total : 12 files,  1093 codes, 39 comments, 249 blanks, all 1381 lines
-
-Summary / [Details](details.md) / [Diff Summary](diff.md) / [Diff Details](diff-details.md)
-
-## Languages
-| language | files | code | comment | blank | total |
-| :--- | ---: | ---: | ---: | ---: | ---: |
-| Python | 10 | 584 | 39 | 247 | 870 |
-| JSON | 1 | 509 | 0 | 1 | 510 |
-| Markdown | 1 | 0 | 0 | 1 | 1 |
-
-## Directories
-| path | files | code | comment | blank | total |
-| :--- | ---: | ---: | ---: | ---: | ---: |
-| . | 12 | 1,093 | 39 | 249 | 1,381 |
-| . (Files) | 2 | 43 | 4 | 25 | 72 |
-| mlmodel | 8 | 980 | 27 | 176 | 1,183 |
-| network | 2 | 70 | 8 | 48 | 126 |
-
-Summary / [Details](details.md) / [Diff Summary](diff.md) / [Diff Details](diff-details.md)
