--- .VSCodeCounter/2023-03-27_11-43-01/results.txt
@@ -1,41 +1,0 @@
-Date : 2023-03-27 11:43:01
-Directory : /Users/gbastien/Documents/Research/Projects/VirusInfectionPredictor/vip/
-Total : 12 files,  1093 codes, 39 comments, 249 blanks, all 1381 lines
-
-Languages
-+----------+------------+------------+------------+------------+------------+
-| language | files      | code       | comment    | blank      | total      |
-+----------+------------+------------+------------+------------+------------+
-| Python   |         10 |        584 |         39 |        247 |        870 |
-| JSON     |          1 |        509 |          0 |          1 |        510 |
-| Markdown |          1 |          0 |          0 |          1 |          1 |
-+----------+------------+------------+------------+------------+------------+
-
-Directories
-+--------------------------------------------------------------------------------------------------------+------------+------------+------------+------------+------------+
-| path                                                                                                   | files      | code       | comment    | blank      | total      |
-+--------------------------------------------------------------------------------------------------------+------------+------------+------------+------------+------------+
-| .                                                                                                      |         12 |      1,093 |         39 |        249 |      1,381 |
-| . (Files)                                                                                              |          2 |         43 |          4 |         25 |         72 |
-| mlmodel                                                                                                |          8 |        980 |         27 |        176 |      1,183 |
-| network                                                                                                |          2 |         70 |          8 |         48 |        126 |
-+--------------------------------------------------------------------------------------------------------+------------+------------+------------+------------+------------+
-
-Files
-+--------------------------------------------------------------------------------------------------------+----------+------------+------------+------------+------------+
-| filename                                                                                               | language | code       | comment    | blank      | total      |
-+--------------------------------------------------------------------------------------------------------+----------+------------+------------+------------+------------+
-| /Users/gbastien/Documents/Research/Projects/VirusInfectionPredictor/vip/__init__.py                    | Python   |          0 |          0 |          1 |          1 |
-| /Users/gbastien/Documents/Research/Projects/VirusInfectionPredictor/vip/mlmodel/README.md              | Markdown |          0 |          0 |          1 |          1 |
-| /Users/gbastien/Documents/Research/Projects/VirusInfectionPredictor/vip/mlmodel/__init__.py            | Python   |          0 |          0 |          1 |          1 |
-| /Users/gbastien/Documents/Research/Projects/VirusInfectionPredictor/vip/mlmodel/build_model.ipynb      | JSON     |        509 |          0 |          1 |        510 |
-| /Users/gbastien/Documents/Research/Projects/VirusInfectionPredictor/vip/mlmodel/compute_ml_features.py | Python   |        265 |          9 |         86 |        360 |
-| /Users/gbastien/Documents/Research/Projects/VirusInfectionPredictor/vip/mlmodel/genes_features.py      | Python   |          2 |          1 |         12 |         15 |
-| /Users/gbastien/Documents/Research/Projects/VirusInfectionPredictor/vip/mlmodel/genomes_features.py    | Python   |        189 |         17 |         68 |        274 |
-| /Users/gbastien/Documents/Research/Projects/VirusInfectionPredictor/vip/mlmodel/hyperparameters.py     | Python   |          0 |          0 |          1 |          1 |
-| /Users/gbastien/Documents/Research/Projects/VirusInfectionPredictor/vip/mlmodel/read_sequence.py       | Python   |         15 |          0 |          6 |         21 |
-| /Users/gbastien/Documents/Research/Projects/VirusInfectionPredictor/vip/network/gephi_input.py         | Python   |          0 |          0 |          1 |          1 |
-| /Users/gbastien/Documents/Research/Projects/VirusInfectionPredictor/vip/network/properties.py          | Python   |         70 |          8 |         47 |        125 |
-| /Users/gbastien/Documents/Research/Projects/VirusInfectionPredictor/vip/predict_interactions.py        | Python   |         43 |          4 |         24 |         71 |
-| Total                                                                                                  |          |      1,093 |         39 |        249 |      1,381 |
-+--------------------------------------------------------------------------------------------------------+----------+------------+------------+------------+------------+
