--- VirusInfectionPredictor_project.egg-info/PKG-INFO
@@ -1,4 +1,0 @@
-Metadata-Version: 2.1
-Name: VirusInfectionPredictor-project
-Version: 0.1.0
-License-File: LICENSE
