+++ pyproject.toml
@@ -1,0 +1,28 @@
+[tool.poetry]
+name = "virushostinteractionpredictor"
+version = "0.1.0"
+description = ""
+packages = [
+    { include = "vhip", from = "src" },
+]
+authors = ["Eric Bastien <gbastien@umich.edu>"]
+license = "MIT"
+readme = "README.md"
+
+[tool.poetry.dependencies]
+python = "^3.9"
+numpy = "^1.25.2"
+pandas = {version = "^2.0.3", extras = ["computation"]}
+scikit-learn = "^1.2.2"
+biopython = "^1.78"
+skops = "^0.9.0"
+matplotlib = "^3.8.2"
+seaborn = "^0.13.0"
+
+
+[tool.poetry.group.dev.dependencies]
+ipykernel = "^6.26.0"
+
+[build-system]
+requires = ["poetry-core"]
+build-backend = "poetry.core.masonry.api"
