+++ src/vhip/mlmodel/genes_features.py
@@ -1,0 +1,14 @@
+
+
+
+
+
+# calculate codon bias
+def codon_bias(gene):
+    pass
+
+
+
+
+
+
