+++ tests/test_d2Distance.py
@@ -1,10 +1,11 @@
-from vip.mlmodel.genomes_features import d2Distance, KmerProfile
 import numpy as np
+from vhip.mlmodel.genomes_features import KmerProfile, d2Distance
+
 
 def test_d2Distance_init():
     # test 1 (distance value is meaningful)
-    seq1 = 'ATCCTGAGTA'
-    seq2 = 'CCAGGCCTGA'
+    seq1 = "ATCCTGAGTA"
+    seq2 = "CCAGGCCTGA"
 
     seq1_profile = KmerProfile(seq1, 6)
     seq1_profile.generate_profile()
@@ -13,12 +14,12 @@ def test_d2Distance_init():
 
     test = d2Distance(seq1_profile, seq2_profile)
     test.distance()
-    dist = 0.4997 # distance for the two small sequences above
+    dist = 0.4997  # distance for the two small sequences above
     assert np.round(test.dist, 4) == dist
 
     # test 2 (sequences are the same so distance is 0)
-    seq1 = 'ATTCCTGGAGTGACCGTGATGA'
-    seq2 = 'ATTCCTGGAGTGACCGTGATGA'
+    seq1 = "ATTCCTGGAGTGACCGTGATGA"
+    seq2 = "ATTCCTGGAGTGACCGTGATGA"
 
     seq1_profile = KmerProfile(seq1, 3)
     seq1_profile.generate_profile()
@@ -27,12 +28,12 @@ def test_d2Distance_init():
 
     test = d2Distance(seq1_profile, seq2_profile)
     test.distance()
-    dist = 0 # distance is 0 since sequences are completely similar
+    dist = 0  # distance is 0 since sequences are completely similar
     assert test.dist == dist
 
     # test 3 (edge case where k used for the sequences do not match)
-    seq1 = 'ATCCTGAGTA'
-    seq2 = 'CCAGGCCTGA'
+    seq1 = "ATCCTGAGTA"
+    seq2 = "CCAGGCCTGA"
 
     seq1_profile = KmerProfile(seq1, 3)
     seq1_profile.generate_profile()
@@ -42,6 +43,3 @@ def test_d2Distance_init():
     test = d2Distance(seq1_profile, seq2_profile)
     test.distance()
     assert test.dist == None
-
-
-
