+++ tests/test_read_sequence.py
@@ -1,6 +1,6 @@
-from vip.mlmodel.read_sequence import read_sequence
+from vhip.mlmodel.read_sequence import read_sequence
 
-seq = '''AGTACTTGTTGATGCTGATGCACTAGTTGATTCAGATGTGCTCGTACTTGTTGATTCAGACGCACTTGTG
+seq = """AGTACTTGTTGATGCTGATGCACTAGTTGATTCAGATGTGCTCGTACTTGTTGATTCAGACGCACTTGTG
 CTCGCTGAAGTACTATTAGATGTAGACGTGCTTGCGCTTATCGATTCAGAAGTACTTGTACTTTCTGAAC
 TACTCTCTGAAGTCGAATTACTTAATGAAGTACTTTCACTATTTTTATCTCTTGCTGATTCGCTTTCTGA
 TAATGATGCACTGTTTACTGAATTACTCGTTGAAGTACTTGCTGATTGTAAACGAGATAGTGAATCACTG
@@ -17,18 +17,16 @@ TCTGAAAGTCGCTTAGAATTACTTTGTTGCTCTGAAAGTCGCTTAGAATTACTTTGTTGCTCTGAAAGTC
 GCTTAGAAGTGCTTGCTTGCTCTGAAAGACGTTTAGAATTACTTTGTTGTTCTGAAGCTTTCTTAGAAGT
 GCTTTGTTGTTCAGAGGCTCTCGATTGACTTTGAGAATCCGCAATACTTTGGCTTGCAGATTTCGATGCA
 ACAGATGGGTCTTCACCATGAATTACTGGTGCAGCGCCACCACGGCCTATAAAGAAGTTTCTATTTGAAT
-CTGTGGAACTCGCAAATGGAACAAAGTCCCAAGTTGCATTATCAGCTATAATATTTGCGCTCCATTTTGT'''
+CTGTGGAACTCGCAAATGGAACAAAGTCCCAAGTTGCATTATCAGCTATAATATTTGCGCTCCATTTTGT"""
 
-cleaned_seq = seq.replace('\n', '') #remove \n created by long string
+cleaned_seq = seq.replace("\n", "")  # remove \n created by long string
+
 
 def test_read_sequence():
     # test 1 - check length of sequences match
-    filename = 'tests/datatests/test_sequence.fasta'
+    filename = "tests/datatests/test_sequence.fasta"
     res = read_sequence(filename)
     assert len(res) == len(cleaned_seq)
 
     # test 2 - check that sequences do match
     assert res == cleaned_seq
-
-
-
