--- vip/mlmodel/genes_features.py
@@ -1,14 +1,0 @@
-
-
-
-
-
-# calculate codon bias
-def codon_bias(gene):
-    pass
-
-
-
-
-
-
